rename typing module into typing_context
